// api/generated/model/trainingByDateDetailsDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { EnrichedExerciseDto } from './enrichedExerciseDto';
import type { PlanDayChooseDto } from './planDayChooseDto';

export interface TrainingByDateDetailsDto {
  /** @nullable */
  _id?: string | null;
  /** @nullable */
  type?: string | null;
  createdAt?: string;
  planDay?: PlanDayChooseDto;
  /** @nullable */
  gym?: string | null;
  /** @nullable */
  exercises?: EnrichedExerciseDto[] | null;
}